moved code from AuditService to new handlers
